Refactor menu event handling code [2/4]
Another refactoring CL for menu element code. Previously, the command
invoker logic was split (and roughly duplicated) between
HTMLButtonElement and HTMLMenuItemElement. Since both can be command
invokers, it makes sense to move the command management stuff to the
common ancestor class, HTMLElement. This CL does that, and de-dupes
the code out of both subclasses. There is now just one base class
method that is overridden in sub-classes that can be command
invokers: CanBeCommandInvoker().

There was also a weird combination of popover activation command
code, which belongs on HTMLElement because any element can be a
popover, and menu activation command code, which belongs on the
HTMLMenuItemElement class, because only that class handles menu
commands.

This touches a few outside classes like ax_object, which get simpler
now that they don't have to deal with each case of a command invoker.

There is one behavior change here: now, if a menuitem is *both*
checkable *and* a sub-menu invoker, the sub-menu wins, and the
item does not do checked/unchecked behaviors.

I also added a test that seemed missing: make sure only buttons and
menuitems can be command invokers. Even though <menuitem> isn't
shipped, this test should pass because it explicitly skips <menuitem>,
and anyway menuitem isn't included in the HTML5_ELEMENTS list.

Bug: 406566432
Change-Id: I281eda0eda4b0dced3ad38bd09a488f2c6b1ddf4